pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "demo-process-observability"
version = "0.0.0"
description = "Demo repo for process observability"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Demo", email = "demo@example.com" }]
dependencies = [
  "pyyaml>=6.0.0",
  "pydantic>=2",
  "rich>=13.0.0",
  "openai>=1.40.0",
  "python-dotenv>=1.0.0",
  "streamlit>=1.30.0",
  "requests>=2.31.0",
  "google-api-python-client>=2.111.0",
  "google-auth>=2.23.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.urls]
Homepage = "https://example.com"

[project.scripts]
demo = "demo.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/demo"]
